+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @brains/ui-library
 
+## 0.2.0-alpha.406
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/contracts@0.2.0-alpha.406
+  - @brains/operator-view-react@0.2.0-alpha.406
+  - @brains/utils@0.2.0-alpha.406
+  - @brains/plugins@0.2.0-alpha.406
+
 ## 0.2.0-alpha.405
 
 ### Patch Changes