+++ plugins/atproto-registry/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @brains/atproto-registry
 
+## 0.2.0-alpha.406
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/atproto-contracts@0.2.0-alpha.406
+  - @brains/utils@0.2.0-alpha.406
+  - @brains/plugins@0.2.0-alpha.406
+
 ## 0.2.0-alpha.405
 
 ### Patch Changes