+++ plugins/directory-sync/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @brains/directory-sync
 
+## 0.2.0-alpha.406
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/content-formatters@0.2.0-alpha.406
+  - @brains/contracts@0.2.0-alpha.406
+  - @brains/image@0.2.0-alpha.406
+  - @brains/utils@0.2.0-alpha.406
+  - @brains/entity-service@0.2.0-alpha.406
+  - @brains/plugins@0.2.0-alpha.406
+
 ## 0.2.0-alpha.405
 
 ### Patch Changes