+++ shared/operation-context/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @brains/operation-context
 
+## 0.2.0-alpha.406
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/contracts@0.2.0-alpha.406
+
 ## 0.2.0-alpha.405
 
 ### Patch Changes