+++ shared/operator-view-react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @brains/operator-view-react
 
+## 0.2.0-alpha.406
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/app-ui-react@0.2.0-alpha.406
+  - @brains/plugins@0.2.0-alpha.406
+
 ## 0.2.0-alpha.405
 
 ### Patch Changes